+++ src/reppi/dictionary/frozen/frozen.py
@@ -161,7 +161,7 @@ class IncrementalFrozenDictionary:
         -------
         self
         """
-        X = np.asarray(X, dtype=float)
+        X = np.asarray(X, dtype=np.float32)
 
         learner = self.base_learner_class(**self.base_learner_kwargs)
         logger.info("Fitting base dictionary with %d samples and %d features using %s", X.shape[1], X.shape[0], self.base_learner_class.__name__)
@@ -257,7 +257,7 @@ class IncrementalFrozenDictionary:
                 f"class_label {class_label} has already been used."
             )
 
-        X = np.asarray(X, dtype=float)
+        X = np.asarray(X, dtype=np.float32)
 
         kwargs = {**self.residual_learner_kwargs, **(learner_kwargs_override or {})}
         logger.info("Adding class %d with %d samples and %d features using %s", class_label, X.shape[1], X.shape[0], self.residual_learner_class.__name__)
